+++ src/project/core/config.py
@@ -6,18 +6,18 @@ load_dotenv()
 
 
 class Settings(BaseSettings):
-    ORIGINS: str
-    ROOT_PATH: str
-    ENV: str
-    LOG_LEVEL: str
+    ORIGINS: str = ""
+    ROOT_PATH: str = ""
+    ENV: str = ""
+    LOG_LEVEL: str = ""
 
-    POSTGRES_SCHEMA: str
-    POSTGRES_HOST: str
-    POSTGRES_DB: str
-    POSTGRES_PORT: int
-    POSTGRES_USER: SecretStr
-    POSTGRES_PASSWORD: SecretStr
-    POSTGRES_RECONNECT_INTERVAL_SEC: int
+    POSTGRES_SCHEMA: str = "public"
+    POSTGRES_HOST: str = "localhost"
+    POSTGRES_DB: str = "postgres"
+    POSTGRES_PORT: int = 5433
+    POSTGRES_USER: SecretStr = "postgres"
+    POSTGRES_PASSWORD: SecretStr = "postgres"
+    POSTGRES_RECONNECT_INTERVAL_SEC: int = 1
 
     @property
     def postgres_url(self) -> str: