+++ src/project/infrastructure/postgres/models.py
@@ -1,7 +1,19 @@
+import datetime
+
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from src.project.infrastructure.postgres.database import Base
+
+
+class Users(Base):
+    __tablename__ = 'users'
+
+    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
+    name: Mapped[str] = mapped_column(nullable=False)
+    email: Mapped[str] = mapped_column(nullable=False, unique=True)
+    role: Mapped[str] = mapped_column(nullable=False)
+    password_hash: Mapped[str] = mapped_column(nullable=False)
 
 
 class Recipe(Base):
@@ -18,8 +30,8 @@ class Recipe(Base):
 class RecipeProduct(Base):
     __tablename__ = "recipe_product"
 
-    id_recipe: Mapped[int] = mapped_column(ForeignKey("recipes.id"), primary_key=True)
-    id_product: Mapped[int] = mapped_column(ForeignKey("products.id"), primary_key=True)
+    id_recipe: Mapped[int] = mapped_column(ForeignKey("recipe.id"), primary_key=True)
+    id_product: Mapped[int] = mapped_column(ForeignKey("product.id"), primary_key=True)
 
 
 class Product(Base):
@@ -30,6 +42,7 @@ class Product(Base):
     cost: Mapped[int] = mapped_column(nullable=False)
 
     recipes = relationship("Recipe", secondary="recipe_product", back_populates="products")
+    batch_product = relationship("BatchProduct", back_populates="product")  # Исправлено
 
 
 class Dish(Base):
@@ -55,6 +68,7 @@ class Order(Base):
 
     waiter = relationship("Waiter", back_populates="orders")
     customer = relationship("Customer", back_populates="orders")
+    orders_dish_cook = relationship("OrdersDishCook", back_populates="order")  # Исправлено
 
 
 class Cook(Base):
@@ -103,3 +117,49 @@ class OrdersDishCook(Base):
     order = relationship("Order", back_populates="orders_dish_cook")
     dish = relationship("Dish")
     cook = relationship("Cook")
+
+
+class Warehouse(Base):
+    __tablename__ = "warehouse"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    location: Mapped[str] = mapped_column(nullable=False)
+    how_full: Mapped[float] = mapped_column(nullable=False)
+
+    batch = relationship("Batch", back_populates="warehouse")
+
+
+class Supplier(Base):
+    __tablename__ = "supplier"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(nullable=False)
+    cost: Mapped[int] = mapped_column(nullable=False)
+    rating: Mapped[float] = mapped_column(nullable=False)
+
+    batch = relationship("Batch", back_populates="supplier")
+
+
+class Batch(Base):
+    __tablename__ = "batch"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    id_supplier: Mapped[int] = mapped_column(ForeignKey("supplier.id"), nullable=False)
+    id_warehouse: Mapped[int] = mapped_column(ForeignKey("warehouse.id"), nullable=False)
+    total_cost: Mapped[int] = mapped_column(nullable=False)
+
+    supplier = relationship("Supplier", back_populates="batch")  # Исправлено
+    warehouse = relationship("Warehouse", back_populates="batch")  # Исправлено
+    batch_product = relationship("BatchProduct", back_populates="batch")  # Исправлено
+
+
+class BatchProduct(Base):
+    __tablename__ = "batch_product"
+
+    id_batch: Mapped[int] = mapped_column(ForeignKey("batch.id"), primary_key=True)
+    id_product: Mapped[int] = mapped_column(ForeignKey("product.id"), primary_key=True)
+    amount: Mapped[int] = mapped_column(nullable=False)
+    expiration_date: Mapped[datetime.date] = mapped_column(nullable=False)
+
+    batch = relationship("Batch", back_populates="batch_product")
+    product = relationship("Product", back_populates="batch_product")